making the real scale

+++ simulator/src/common/app.cxx
@@ -33,14 +33,14 @@ void App::createScene() {
   this->shipView = new ShipView(this->m_SceneMgr, this->cameraManager, "ship");
   MathUtilities::scaleFactor = 0.01;
   //this->planets.push_back(new PlanetView(this->m_SceneMgr, "sol", "sun.mesh", "", Ogre::Vector3(0, 0, 0), Ogre::Real(10), Ogre::Real(40), Ogre::Radian(0.01)));
-  this->planets.push_back(new PlanetView(this->m_SceneMgr,  "mercurio",  "mercury.mesh",  "", this->shipView->getShip(), Ogre::Vector3(57909175000, 0, 0),   Ogre::Real(3.302e23), Ogre::Real(240000000),    Ogre::Radian(0.01)));
-  this->planets.push_back(new PlanetView(this->m_SceneMgr,  "venus",     "Venus.mesh",    "", this->shipView->getShip(), Ogre::Vector3(108208930000, 0, 0),  Ogre::Real(4.8690e24), Ogre::Real(600000000),   Ogre::Radian(0.01)));
-  this->planets.push_back(new PlanetView(this->m_SceneMgr,  "tierra",    "earth.mesh",    "", this->shipView->getShip(), Ogre::Vector3(149597870000, 0, 0),  Ogre::Real(5.9742e24), Ogre::Real(630000000),   Ogre::Radian(0.01)));
-  this->planets.push_back(new PlanetView(this->m_SceneMgr,  "marte",     "Mars.mesh",     "", this->shipView->getShip(), Ogre::Vector3(227936640000, 0, 0),  Ogre::Real(6.4191e23), Ogre::Real(340000000),   Ogre::Radian(0.01)));
-  this->planets.push_back(new PlanetView(this->m_SceneMgr,  "jupiter",   "jupiter.mesh",  "", this->shipView->getShip(), Ogre::Vector3(778412010000, 0, 0),  Ogre::Real(1.8987e27), Ogre::Real(7500000000),  Ogre::Radian(0.01)));
-  this->planets.push_back(new PlanetView(this->m_SceneMgr,  "saturno",   "Saturn.mesh",   "", this->shipView->getShip(), Ogre::Vector3(1426725400000, 0, 0), Ogre::Real(5.6851e26), Ogre::Real(6000000000),  Ogre::Radian(0.01)));
-  this->planets.push_back(new PlanetView(this->m_SceneMgr,  "urano",     "Urano.mesh",    "", this->shipView->getShip(), Ogre::Vector3(2870972200000, 0, 0), Ogre::Real(8.6849e25), Ogre::Real(2530000000),  Ogre::Radian(0.01)));
-  this->planets.push_back(new PlanetView(this->m_SceneMgr,  "neptuno",   "Neptuno.mesh",  "", this->shipView->getShip(), Ogre::Vector3(4498252900000, 0, 0), Ogre::Real(1.0244e26), Ogre::Real(2460000000),  Ogre::Radian(0.01)));
+  this->planets.push_back(new PlanetView(this->m_SceneMgr,  "mercurio",  "mercury.mesh",  "", this->shipView->getShip(), Ogre::Vector3(57909175000, 0, 0),   Ogre::Real(3.302e23), Ogre::Real(24000000),    Ogre::Radian(0.01)));
+  this->planets.push_back(new PlanetView(this->m_SceneMgr,  "venus",     "Venus.mesh",    "", this->shipView->getShip(), Ogre::Vector3(108208930000, 0, 0),  Ogre::Real(4.8690e24), Ogre::Real(60000000),   Ogre::Radian(0.01)));
+  this->planets.push_back(new PlanetView(this->m_SceneMgr,  "tierra",    "earth.mesh",    "", this->shipView->getShip(), Ogre::Vector3(149597870000, 0, 0),  Ogre::Real(5.9742e24), Ogre::Real(63000000),   Ogre::Radian(0.01)));
+  this->planets.push_back(new PlanetView(this->m_SceneMgr,  "marte",     "Mars.mesh",     "", this->shipView->getShip(), Ogre::Vector3(227936640000, 0, 0),  Ogre::Real(6.4191e23), Ogre::Real(34000000),   Ogre::Radian(0.01)));
+  this->planets.push_back(new PlanetView(this->m_SceneMgr,  "jupiter",   "jupiter.mesh",  "", this->shipView->getShip(), Ogre::Vector3(778412010000, 0, 0),  Ogre::Real(1.8987e27), Ogre::Real(750000000),  Ogre::Radian(0.01)));
+  this->planets.push_back(new PlanetView(this->m_SceneMgr,  "saturno",   "Saturn.mesh",   "", this->shipView->getShip(), Ogre::Vector3(1426725400000, 0, 0), Ogre::Real(5.6851e26), Ogre::Real(600000000),  Ogre::Radian(0.01)));
+  this->planets.push_back(new PlanetView(this->m_SceneMgr,  "urano",     "Urano.mesh",    "", this->shipView->getShip(), Ogre::Vector3(2870972200000, 0, 0), Ogre::Real(8.6849e25), Ogre::Real(253000000),  Ogre::Radian(0.01)));
+  this->planets.push_back(new PlanetView(this->m_SceneMgr,  "neptuno",   "Neptuno.mesh",  "", this->shipView->getShip(), Ogre::Vector3(4498252900000, 0, 0), Ogre::Real(1.0244e26), Ogre::Real(246000000),  Ogre::Radian(0.01)));
 }
 
 bool App::frameRenderingQueued(const Ogre::FrameEvent& evt) {
